Add enemy pathing AI
Currently, the pathing AI requires that a `Navigation2D` node is set up
with a `NavigationPolygonInstance` which has all colliders on the map
sliced out of it. The removal of the colliders from the navigation mesh
does not happen automatically, and if it is not done, navigation points
can be placed inside of a collision shape, making them inaccessible.

The `Player` and `Navigation2D` nodes must register themselves with the
`Scene` singleton for the AI to work.